cdk/app/requirements.txt
streamlit
pandas
numpy
requests
Pillow
matplotlib
requests
boto3
botocore
tensorflow
opencv-python
PyYAML